@@ -337,7 +337,7 @@ func (a *Awaitility) WaitUntiltMetricHasValue(t *testing.T, family string, expec
 	t.Logf("waiting for metric '%s{%v}' to reach '%v'", family, labels, expectedValue)
 	var value float64
 	err := wait.Poll(a.RetryInterval, a.Timeout, func() (done bool, err error) {
-		value, err := metrics.GetMetricValue(a.RestConfig, a.MetricsURL, family, labels)
+		value, err = metrics.GetMetricValue(a.RestConfig, a.MetricsURL, family, labels)
 		// if error occurred, ignore and return `false` to keep waiting (may be due to endpoint temporarily unavailable)
 		// unless the expected value is `0`, in which case the metric is bot exposed (value==0 and err!= nil), but it's fine too.
 		return (value == expectedValue && err == nil) || (expectedValue == 0 && value == 0), nil
